Fixed a bug in handling of weak pointers by the mark-compact garbage collector.
When threading pointers in 'updateForwardPointersForMarkCompact', we
call 'clearIfWeakAndUnmarkedForMarkCompact' and also compress a
sequence of unmarked objects into one (word8) vector.  There is a bug
here, because 'clearIfWeakAndUnmarkedForMarkCompact' inspects the
header of the object pointer to determine if it is still live.
However, if the object pointer is unmarked and compressed into a
(word8) vector, then 'clearIfWeakAndUnmarkedForMarkCompact' might see
the counter word of the compressed (word8) vector, because
the normal object layout:
  header :: <>(non heap-pointers)* :: (heap pointers)* :::: ...
becomes
  counter :: <>length :: header :: (word8)*
where <> denotes the location pointed to by the object pointer of the
weak pointer.

While I think it would be correct to always clear the weak pointer if
the inspected header is 0 (equal to the newly written counter word),
it seems cleaner to process all the weak pointers before threading
pointers.  To do so, 'dfsMarkByMode' takes a 'bool shouldLinkWeaks';
when true (as in the mark-compact marking phase), during the dfs, all
encountered weaks are linked.  The
'clearIfWeakAndUnmarkedForMarkCompact' function is replaced by a
'updateWeaksForMarkCompact' function, which clears weak pointers whose
object pointer points to an unmarked object.  This function
constitutes a new phase that runs after marking and before pointer
threading.

This should also be more efficient, since the overhead to handle weak
pointers is now proportional to the number of (live) weak pointers in
the heap, rather than proportional to the number of (live) objects in
the heap.

Thanks to Sean McLaughlin for the bug report and Florian Weimer for
the initial diagnosis.



SVN r6183