Fixed bug in type inference of flexible records.  This would later cause the compiler to raise the TypeError exception when type checking the XML program after the defunctorize pass.
The following program exhibits the bug:

fun f (a, b) = (#value a) < (#value b)

fun g (array1 (* : {value:nat, offset:'a1} array *), 
       array2 (* : {value:nat, offset:'a2} array *),
       record1 : {value:nat, offset:'b},
       record2 : {value:nat, offset:'c}) =
   let
      val left (* : {value:nat, ...} *) = Array.sub (array1, 0)
      val right (* : {value:nat, ...} *) = Array.sub (array1, 1)
      val small (* : {value:nat, ...} *) = 
         if f (left, right)
           then Array.sub (array2, 0)
           else Array.sub (array2, 1)
   in
      (f (record1, small),
       f (small, record2))
   end

After elaborating "fun f (a, b) = ...", f is assigned the type scheme:
  {value:int, gen1:...} * {value:int, gen2:...} -> bool
where gen1 and gen2 denote (distinct) generalizable flex records, each
with their own spine.

After elaborating "val left = ..." and "val right = ...", left and
right are assigned (the same) unification variable, which is otherwise
unconstrained.

In elaborating "f (left, right)", we instantiate the type scheme of f,
realizing the two generalizable flex records as flex records.  We then
unify the flex records with left and with right.  Since left and right
have the same unification variable, this effectively unifies the two
flex records, which unifies the two spines.  This is correct, as we
now know that the spines of the two record arguments to f must be the
same.  However, the generalizable fields may be generalized to
different types in the two arguments.  But, Spine.unify causes the two
spines for the record arguments to be equal according to Spine.equals.
This causes problems for computing the type arguments to an
application of f, because we use Spine.equals to find the flexible
record type that was derived from the generalizeable flex record at
scheme instantiation.  Nothing goes wrong with "f (left, right)",
because left and right have the same type, so the instantiation of the
offset field type will be the same.  However, with "f (record1,
small)" the instantiations of the offset field type will be different
('b for record1, and 'a2 for small).  The spine unification caused the
lookup of the type argument from the generalizable fields of the first
record to see the types of the second record, leading to the XML type
error:
  Type error: actual and formal not of same type
  actual: ('a_4092 * word32) * (word32 * word32)
  formal: (word32 * word32) * (word32 * word32)

The fix was to give Spine.t's an identity separate from the
DisjointSet.t used to ensure that spines had the same sets of fields.
The spine identity is now used for Spine.equals in the lookup.



SVN r6229