Fixed bug in the FFI visible representation of Int16.in ref (and references to other primitive types smaller than 32-bits) on big-endian platforms.
Added special logic to PackedRepresentation.TupleRep.make to handle
the case of a boxed tuple that is entirely comprised of primitive
types.  (This is a strict super-set of 'Int16.int ref' and related types.)
On a big-endian platform, sub-word32 components must be at the
low byte offset (but high bit offset) of the containing word32.  This
is fairly easy to handle by just reversing the order of the sub-word32
components and padding at the low bits on a big endian platform.

We could be more selective and only special case single element tuples
with a mutable sub-word32 component, which more closely approximates
the 'Int16.int ref' and related types that may appear in the FFI.

The representation and operations should be identical on a
little-endian system before and after this change.  Clearly, things
change a little on a big-endian system.  Effectively, initializing a
special-cased tuple (potentially) requires more work.  

For example, on a big-endian plaform, an (unflattened) 'Int16.int ref'
used to be represented like:
    opt_7 = Normal {hasIdentity = true, ty = [Word8, Bits24]}
and initialized like:
    x_0: Word8
    thing8_0: Objptr (opt_7) = Object {header = 0xF, size = 8}
    x_2: Bits32 = WordU8_extdToWord32 (x_0)
    x_1: [Word8, Bits24] = x_2
    OW32 (thing8_0, 0): [Word8, Bits24] = x_1
and updated like:
    OW8 (thing8_0, 3): Word8 = x_19
and derefed like:
    x_20: Word8 = OW8 (thing8_0, 3): Word8

Now, on a big-endian platform, it is represented like
    opt_7 = Normal {hasIdentity = true, ty = [Bits24, Word8]}    
and initialized like:
    x_0: Word8
    thing8_0: Objptr (opt_7) = Object {header = 0xF, size = 8}
    x_3: Bits32 = WordU8_extdToWord32 (x_0)
    x_2: Bits32 = Word32_lshift (x_3, 0x18: Word32)
    x_1: [Bits24, Word8] = x_2
    OW32 (thing8_0, 0): [Bits24, Word8] = x_1
and updated like:
    x_23: Word8
    OW8 (thing8_0, 0): Word8 = x_23
and derefed like:
    x_24: Word8 = OW8 (thing8_0, 0): Word8

Note that we pay an extra lshift in the initialization to get the
Word8 to the high-bits of the containing Bits32.  (For an '(Int8.int *
Int8.int) ref', we would need shift both Word8s to the high bits, and
do an orb; on a little endian system,we only need to shift one Word8,
and do an orb.)  On the other hand, accessing the Word8 can be done
with a zero-byte offset.  Unfortunately, all of our big-endian systems
are RISC, so I don't think we save any instruction space by having a
zero-byte offset.  (I think on x86 and amd64, a zero-byte offset can
be done with a smaller instruction encoding; though, they are
little-endian.)

Also simplified the logic of PackedRepresentation.TupleRep.make.  It
turns out that we can determine whether to box and whether to
padToPrim just by examining the list of tuple component
representations.  Hence, we need fewer redirections and suspensions.



SVN r6267