Optimized min, max and compareReal.
The idea is to "delay" treatment of NaNs, which usually are very rare, by
relying on the property that ordering predicates on reals return false
when at least one of the reals being compared is NaN.

The optimization of min and max should not cause any regressions compared
to the previous implementation.  However, the new version of compareReal
no longer makes a call to C, which means that on the x86, MLton may keep
reals in registers, and this may cause some corner cases to be evaluated
differently, due to the use of 80-bit reals by the x86 FPU.

Note that all of min, max and compareReal could be implemented using just
a single floating point comparison on the x86.


SVN r6291