Reverted a change made in revision 6046.
Tycon.newString and Tycon.fromString have the same effect except that
Tycon.newString leaves the printName of the tycon unset while
Tycon.fromString sets the given string as the printName for type tycon.

In revision 6046, the code was changed to use Tycon.newString.  An effect
of that change was that the names of many (more?) tycons in compiler
output (e.g. output produced by show-basis) included a generated suffix.
Although such names are more precise in some sense, the generated suffices
make the tycon names less readable to a programmer.

This commit reverts the code to use Tycon.fromString.  Regression tests
seem to pass and the code from the bug report that was fixed by revision
6046 also seems to compile correctly.


SVN r6295