Extended processDefUse to save types of bindings.
To make the types accessible at the point processDefUse is called, newUses
was extended to save the types (as a list of range values) along with the
defUses of variables.

To make the types more readable to programmers, a couple of changes were
introduced.  setTyconNames was changed to avoid adding the "?." prefixes
(noise) to tycon names when it is called with a top-level scope.  To make
types defined inside functor arguments and bodies recognizable outside the
functor, FunctorClosure was extended to include the formal parameter Strid
and a module path prefix.


SVN r6333