Slightly generalized MLton's overload mechanism and improved checking of overload definitions.
An overload now carries a (type) scheme rather than a type.  This allows
polymorphic values to be overloaded.  For example, an overloaded sub could
be defined as follows:

  _overload 2 sub : 'a * int -> 'b
  as        List.nth
  and      Array.sub
  and     Vector.sub
  and  CharArray.sub
  and CharVector.sub

The list of variants for an overload is now allowed to include other
overload definitions.  This allows overload definitions to be (re)defined
incrementally.  For example, the overload for + could be extended with a
new function as follows:

  _overload 2 + : 'a * 'a -> 'a as + and My.plus

When elaborating an overload definition, it will now be checked that all
the given variants unify with the scheme given for the overload.  For
example, the invalid definition:

  _overload 2 funny : int -> int as Real.+ and CharVector.sub

now gives the following errors:

  Error: [...]
    Variant does not unify with overload.
      overload: [int] -> [int]
      variant:  [real * real] -> [real]
      in: _overload 2 funny: (int -> int) as Real.+ and CharVector.sub
  Error: [...]
    Variant does not unify with overload.
      overload: [int] -> [int]
      variant:  [string * int] -> [char]
      in: _overload 2 funny: (int -> int) as Real.+ and CharVector.sub


SVN r6335