Support processor-local allocation.
The GC_heap structure now defines the global nursery, while each
processor's GC_state tracks the per-processor allocation area.
Allocation from a processor-local area occurs just as for the single-
processor version of MLton.  When processor-local space is exhausted,
the processor attempts to take a chunk from the global space using an
atomic compare-and-swap operation.  If no global space is available,
the processor signals that a collection is necessary and waits for the
remaining processors to catch up.

One complicating factor is that we must ensure that the unused portion
of a processor's allocation area can always be filled.  To do so, we
add a small amount of space to each processor-local allocation area that
ensures that the area is large enough to be filled.  This additional
space is known as the "bonus slop" in the code.


SVN r6441