Functor to abstract a representation of a type.
A number of the system types (e.g., Posix.FileSys.file_desc) are not
abstracted, due to the fact that they are defined as follows:

  structure C_Int = struct open Int32 type t = int end
  structure C_Fd = C_Int

  structure PosixIO = struct
    ...
    type file_desc = C_Fd.t
    ...
  end

  signature BASIS_EXTRA = sig
    ...
  end
  (* Types referenced in signatures by structure name *)
  where Posix.IO.file_desc = Posix.IO.file_desc

This reveals that Posix.IO.file_desc is equal to Int32.int.

This commit introduces a functor to abstract the representation of a
type, to be used to elimate the type leakage described above.


SVN r6536