Changed dominators to use an implementation of the algorithm described in
  A Simple, Fast Dominance Algorithm.
  Keith Cooper and Timothy Harvey and Ken Kennedy.
  Software Practice and Experience, 2001.
  http://citeseer.ist.psu.edu/cooper01simple.html
  http://www.cs.rice.edu/~keith/EMBED/dom.pdf

This implementation replaced the previous implementation based on the
Lengauer/Tarjan algorithm, as described on p. 185-191 of Muchnick's
"Advanced Compiler Design and Implementation", and appears to run in less
than half the time on a self compile and took about half the amount of
code to implement.  Much of the speed up probably comes from less memory
being used and the use of (faster) arrays rather than (slower) property
lists.


SVN r6610