Output PIC assembly code. This requires accessing global symbols using RIP-relative addressing. RIP-relative addressing also results in about 2% smaller executables, so this code is used for non-libraries as well.
The patch locates all instances of:
	PDC Address: movq $name, %rax
	PDC Data Access: movq name, %rax
and changes them to:
	PIC Internal Address: leaq name(%rip),%rax
	PIC Internal Data Access: movq name(%rip),%rax

To effect this change, all cases where a mov was explicitly used with an
immediate label were changed to a lea of the label's memloc. Additionally,
the register-allocation was changed to load symbols relative to RIP. 

Unfortunately, it is not possible to write: movq foo(%rip,%rdi,8),%rax.
This means that the register allocation has to watch for the case where 
there is an index register and if so, break the mov down into:
	leaq foo(%rip),%rax
	movq (%rax,%rdi,8),%rax
Since this is slightly larger, it is only used for library output where 
PIC code is required. For the normal executable code-path, we output
	movq foo(,%rdi,8),%rax
as before.

Since the generated PIC code uses RIP-relative addressing, it is only able 
to access symbols compiled into the same relocatable text segment. Fortunately,
the amd64 codegen makes all system calls via the basis which is compiled in.
The basis in turn makes calls to other DSO text segments. Since the basis
uses -fPIC, this means calls will succeed.

There remains the case where the FFI is used to _import symbols from external 
libraries directly. This will be fixed in a future patch.



SVN r6677