Added {internal,external} attributes to _address, _symbol, _import, _export. These control the scope of the symbol with respect to the linker.
_export and _symbol alloc count as exporting. 
_import, _address, and _symbol without alloc count as importing.

For both imported and exported symbols, the default attribute is external.

External export means that when compiled into a library, these symbols will
be available. However, within the scope of compiling MLton, these symbols
are still used as INTERNAL because they are within the same text segment and
the compiler can optimize to reach them. In headers, these symbols are
marked as MLLIB_EXPORTED. The definitions of these symbols are marked as
EXPORTED to ensure they are public symbols.

Internally exported symbols will not be available to users of the library,
but are available to code being linked into the library. For example, if the
user _exports "foo" internal, he can use "foo" in his linked-in C code, but
this symbol will not be available to users of the resulting library. In the
emitted header file, these symbols are marked as MLLIB_INTERNAL. Both the
definition and use of these symbols are INTERNAL to keep the symbol private
and also to inform the compiler about the potential to optimize.

Externally imported symbols are more costly to access. Particularly for
shared libraries, it requires performing a symbol table lookup. External is
the default because it is always works, even if slower. It is necessary to
use external import if the symbol is located in another text segment. The
most common example of this is a symbol imported from a shared library.

Internally imported symbols must be linked directly with the ML code. Good
examples of this include the entirity of the MLton runtime which is linked
directly with the generated code. If a user is importing his own symbols
which are linked together with the ML code, _import internal is appropriate.
By making the entire runtime internal, no ML implementation details are
exposed in the resulting shared library.

This patch propogates all the internal/external information to the C codegen.
Because the bytecode codegen uses the same C declarations, it works too.

The amd64 & x86 codegens currently implicitly treat all _imports as
internal.  Wrapper functions should be emitted to proxy external calls. This
has not been a problem so far, because on these architectures the linker
fixes up all symbols in executables. Therefore, wrappers (one implemented)
will only be needed when compiling to a shared library.




SVN r6680