Mark all of the runtime as internal. Otherwise the runtime, c-codegen, and bytecode codegen produce assembler using lookup tables in PIC mode.
Special care was required for the MLTON_CODEGEN_WORDSQUOTREM marked functions.
The declarations should be marked INTERNAL for the C codegen, but nowhere else.
This required splitting the macro into one for the declaration and one for
the implementation (so that the implmenetation can still be elided).

Additionally patch gdtoa to not export its symbols beyond linkage with the
ML code. As this is only appropriate for use with MLton, I put this patch
in a separate file.



SVN r6682