This patch adds preliminary support for the win64 ABI. This is sufficient to compile much of the regression suite with native assembly.
The win64 ABI passes at most 4 parameters in registers.
It uses rcx, rdx, r8, and r9 for integers and xmm0-4 for floats.
Unlike on linux, if the 1st parameter is an integer, xmm0 is not used.
eg:
  void foo(int x, float r, float s, int y, int z) =>
  push z
  r9 = y
  xmm2 = s
  xmm1 = r
  rcx = x

Furthermore, parameters passed in registers still reserve space on the stack.
This "shadow space" is always 32 bytes even if there are less parameters.



SVN r6688