Make implementExceptions robust against an un-shrinked input program.
The implementExceptions pass previously required a shrink of the input
program.  The initial extra value was extracted from the argument to
the Exn_setInitExtra primitive.  However, if the input program was not
shrunk, then the variable might not be bound to a primitive
expression.  (Furthermore, even if it were bound to a primitive
expression, the expression may have free variables, and so could not
be lifted to the beginning of the program.) Similarly, if the input
program was not shrunk, then the Exn_extra and Exn_setExtendExtra
primitives might not be eliminated from the input program.

This commit eliminates the Exn_setInitExtra primitive.  To determine
the extra type, we search the program for the type argument to the
Exn_extra and Exn_setExtendExtra primitives.  The extra type should
correspond to a datatype with a nullary constructor, which is used as
the initial extra value.  (In the absence of the primitives, as likely
happens with a shrunk program compiled with 'Exn.keepHistory false',
the extra type is set to unit.)

SVN r6747