Fix MLton_bug prototype.
MLton_bug (the C function) is the realization of MLton_bug (the
primitive) and Bug (the SSA2 transfer).  Within the compiler MLton_bug
(the primitive) is expected to take an ML string argument.  In
SSA-to-RSSA, Bug (the SSA2 transfer) is translated to a call of
MLton_bug (the C function) with an ML string argument.  However,
MLton_bug (the C function) was expecting a null-terminated string,
which wasn't guaranteed by the compiler.

SVN r6751