Simplifying card/cross map in heap.
The insight is that the old resizeCardMapAndCrossMap function was a
bit misleading, because it copies the contents of the old cross map to
the new cross map.  However, the resizeCardMapAndCrossMap function is
only ever called in the resizeHeap function, which, in turn, is only
ever called in the majorGC function and the GC_unpack function.

After a major GC, the cross map must be cleared, because the major GC
will have collected objects in the old generation, making the cross
map invalid.  So, when resizing the heap (and, hence, resizing or
moving the card/cross maps) after a major collection, it suffices to
clear the resulting cross map, rather than clearing the cross map and
then copying the cleared cross map to its new location.

It seems safe to assume that major garbage collections will be much
more frequent than GC_unpack, so it makes sense to simplify the code
for the common case.

This commit allocates space at the end of the ML object heap for the
card and cross maps.

After creating, swapping, or resizing the heap, it uses
setCardMapAndCrossMap to set gcState.generationalMaps appropriately
(according to gcState.heap.start and gcState.heap.size) and clear the
card and cross maps.  There is no complicated copying of card and
cross maps when heaps are resized.

The only downside is that at a GC_unpack, after resizing the heap
after a minor collection, we clear the cross map and, thus, are forced
to recompute the cross map at the next minor collection.  (Note,
however, that because the runtime does not always use generational
collection, there may in fact have been no valid data in the cross map
and there may be no minor collection before the next major
collection.)

SVN r6776