Use the current heap in growHeap if it satisfies the minSize.
When it is possible to use the current heap, then demand proper growth
from remapHeap and/or createHeap.  That is, demand from remapHeap
and/or createHeap a minimum size that is the average of the desired
size and the current heap size.

There is a slight danger here that, when the current heap can be used,
we do not shrink the heap at all before attempting to allocate a new
heap of the desired size.  There is a risk here that not freeing the
memory from the current heap will prevent us from allocating a heap of
the desired size.

SVN r6783