This changes the symbol tags internal/external to private/public/external. Without this change, it was not possible to correctly access symbols on windows and linux simultaneously. The three categories are as follows:
Private  = defined in this module and not available outside it
Public   = defined in this module and exported for use outside
External = defined in another module
... where module means the final library or executable.

The following would be a syntax error:
  _export "foo" external: ..
  _symbol "foo" external alloc: ...
since it is not possible to define a symbol outside of our own module.

_export private/public behave exactly as you'd expect wrt the library API.

As another example,
  val () = _import "foo" private: ...
means that something that MLton is statically linking against must provide
the symbol "foo" and this symbol is not exported outside of the module.

All of the runtime is private in this sense.

The default tags are:
	_import		external
	_address	external
	_symbol		external
	_symbol alloc	public
	_export		public

The C-codegen should work on all platform/format combinations since we just 
carry the information to the output C. The amd64 codegen is correct for ELF,
and from what I've read it should be correct for darwin-x86_64 and win64 too.

Some code that used to "work" on win64 might be broken by this, in that it 
doesn't link. However, before it could link and silently produce pointers to   
function stubs instead of the actual functions, violating ANSI C.

------------

The only thing that bothers me now is that you need to know at SML compile,
whether you are being statically or dynamically linked with a library. If 
the library is a dll/so/dylib, then external is the correct choice. However,
if it is statically linked, public is technically correct.      

On ELF/darwin, external _import/_symbol is a safe choice in both cases.

Unfortunately, on windows it seems this distinction really matters. You     
generate different assembler for the two cases, and they are incompatible.  
It >is< possible to link against both by assuming the symbol is not
external, only importing functions, and never taking pointers to those
functions. That's how MLton is able to link against gmp both statically and
dynamically at the moment. This works mostly because of the magic stub
libraries MinGW makes. This hack is pretty fragile, though.                   

Therefore, I suppose the right thing to recommend is 'external' for all
library imports (it's also the default). Then, if compiling on windows,
you have to link against the DLL, not statically. 

If someone really does want to link against a library statically on windows,   
then they can still do it using 'public' _import.    



SVN r6785