Fix external symbol with x86-codegen on x86-darwin.
It is a bit difficult to cache the symbol stub label for an external
symbol label, because the x86MLton.prim function is called on each
primitive in the incoming Machine IL program.  On the other hand, it
suffices to emit a symbol stub label for each external symbol
label. Since a _symbol label is almost certainly globalized, there is
unlikely to be any opportunities for sharing.

Use thunk suspensions to avoid allocating labels that will not be
used.

SVN r6818