Fix library regression for MinGW. There were a few problems:   * Symbols start with _, so the dlltool sed expression needed changing   * --globalize doesn't localize the rest, we want --keep-global   * We need to preserve the .drectve section so dlltool works on libarchives   * Include --enable-stdcall-fixups for _address of stdcall fns
In order for the regressions to pass, however, one needs to have a patched
binutils which fixes a binutils/bfd bug. I've worked out a patch with one
of the binutils maintainers, so it will be in the next release. When MLton
is next released the mlton.msi windows installer should include the patch.



SVN r6894