Fix bug in nested _export/_import functions.
In the C codegen, the trampoline for MLton_callFromC terminates when
the 'returnToC' is set to 'TRUE' in the SML/ChunkN code.  However, if
an _export-ed function returned to an _import-ed function returned to
an _export-ed function, then the outer _export-ed function is
executing the trampoline for MLton_callFromC with the 'returnToC'
global set to 'TRUE'.  If the outer _export-ed function terminates
quickly, then it will itself set 'returnToC' to 'TRUE' and exit the
SML/ChunkN code.  However, if continuing the execution of the outer
_export-ed function requires an inter-chunk transfer, then it would
(erroneously) terminate the trampoline (without properly returning
values).

Similarly, when executing as a library, the 'lib_open' trampoline
could erroneously be terminated with 'returnToC' is set to 'TRUE' when
an _export-ed function returns to an _import-ed function returns to
top-level SML code.



SVN r6919