Now "-stop o" and "-stop g" use the name given with "-output ..." so the location and name for the generated files can be customized.
Because "-stop o" and "-stop g" typically generate multiple files, the
suffix (if any) is first stripped from the name given with -output.

For example, this allows one to add "-stop o" to a command to build an
executable with "-output bar/foo.exe" and still get a reasonable
result: object files in bar/foo.*.o.


SVN r6992