Replace use of _const in mlnlffi-lib with a C program which generates a small .sml file for the constants.
It was not a good idea to use _const.  It forces the Basis Library
archive to carry around values that are only used by mlnlffi-lib.
This means that these constants need to be available on every
platform, regardless of whether or not that platform supports libdl
(some may not).


SVN r7036